Refactor Dockerfile to start ElectionWatch API server with main.py; remove deprecated main_server.py file.
