+++ requirements.txt
@@ -1,4 +1,4 @@
-# ElectionWatch Agent API - Cloud Run Requirements
+# ElectionWatch Agent API Service Requirements
 # Minimal dependencies for running the FastAPI agent service
 
 # Core FastAPI service (updated for ADK compatibility)
@@ -8,22 +8,37 @@ pydantic>=2.5.0
 python-multipart>=0.0.6
 python-dotenv>=1.0.0
 
-# Additional dependencies for custom endpoints
-pandas>=2.0.0
-python-json-logger>=2.0.0
+# HTTP clients for examples
+aiohttp>=3.8.0
+websockets>=11.0.0
+
+# Google ADK (optional - will fallback to enhanced coordinator if not available)
+google-adk==1.5.0
+google-cloud-aiplatform[agent_engines]>=1.95.1
+
+# Enhanced coordinator dependencies
+pandas>=1.5.0
+numpy>=1.24.0
+
+# Database connections (for tools)
+neo4j==5.22.0
+pymongo>=4.0.0
+
+# ML dependencies (for tools)
+llama-index==0.10.53
+llama-index-embeddings-huggingface
+transformers>=4.30.0
+torch>=2.0.0
+
+# Optional - for full functionality
+langchain-community==0.2.10
+langgraph==0.1.13
+
+# Additional dependencies for file handling
+python-multipart>=0.0.6
+pydantic>=2.5.0
 
 # Google Cloud dependencies
 google-cloud-core>=2.4.1
 google-auth>=2.23.4
-
-# Updated ADK and AI Platform versions to resolve conflict
-google-adk==1.5.0; python_version>="3.8"
-google-cloud-aiplatform[agent_engines]>=1.95.1; python_version>="3.8"
-
-# Basic data processing (lightweight versions)
-pandas>=2.1.4
-numpy>=1.24.4
-
-# HTTP client for API calls (updated for ADK compatibility)
-aiohttp>=3.9.1
 requests>=2.32.4
